feat: tweaking package and linter rules for typesript test conversion  feat: some of the LLM converted tests
