+++ tests/system/test-api-get-api.ts
@@ -1,5 +1,3 @@
-// @ts-nocheck
-
 import test from 'ava'
 import { startApi } from '../helpers/api.js'
